unify greenplum scripts

+++ bin/move-greenplum.sh
@@ -20,21 +20,31 @@
 # https://gpdb.docs.pivotal.io/6-10/admin_guide/load/topics/g-defining-a-command-based-writable-external-web-table.html
 # and https://gpdb.docs.pivotal.io/6-10/ref_guide/sql_commands/CREATE_EXTERNAL_TABLE.html
 
-# Single argument is the directory where the data to load is located
+# Two arguments:
+# $1: 'fromdb' or 'todb' indicates direction of data movement
+# $2: the directory where the data to load is located
 # This script is invoked by each segment in a segment host
-DIR=$1
-REGEX=".*file([0-9]+).*"
-for file in ${DIR}/*.db; do
-   if [[ ${file} =~ ${REGEX} ]]; then
-      NUMBER=${BASH_REMATCH[1]}
-      # Only allocate some files to each segment
-      SEGMENT=$(( ${NUMBER} % ${GP_SEGMENT_COUNT} ))
-      if [[ "${GP_SEGMENT_ID}" == "${SEGMENT}" ]]; then
-          cat ${file}
-          rm ${file}
+DIR=$2
+PREFIX="file"
+REGEX=".*${PREFIX}([0-9]+).*"
+
+if [ "$1" == "fromdb" ]; then
+   mkdir -p ${DIR} || exit 1
+   #cat </dev/stdin >${DIR}/${PREFIX}${GP_SEGMENT_ID}
+   split -l 500000 -a 3 - ${DIR}/${PREFIX}${GP_SEGMENT_ID} </dev/stdin
+else
+   for file in ${DIR}/*.db; do
+      if [[ ${file} =~ ${REGEX} ]]; then
+         NUMBER=${BASH_REMATCH[1]}
+         # Only allocate some files to each segment
+         SEGMENT=$(( ${NUMBER} % ${GP_SEGMENT_COUNT} ))
+         if [[ "${GP_SEGMENT_ID}" == "${SEGMENT}" ]]; then
+            cat ${file}
+            rm ${file}
+         fi
+      else
+         echo "Unexpected file name"
+         exit 1
       fi
-   else
-       echo "Unexpected file name"
-       exit 1
-   fi
-done
+   done
+fi
